# openapi.yaml
openapi: 3.0.3
info:
  title: SLSEA Real-Time Solar Generation Data API
  version: 0.1.0
  description: |
    REST API for the Sri Lanka Sustainable Energy Authority (SLSEA). Rooftop solar
    meters push generation readings (write path); SLSEA users read the hierarchy,
    installations and generation history, scoped to their jurisdiction (read path).

    ## What is live now

    **Implemented:** `/health` and the whole read path: provinces, districts, grid substations
    and installations (collections with their filters, and members), the installation composite,
    the last-known reading, and one installation's readings collection and reading member.
    Also reading ingestion: `POST /installations/{installation-id}/readings` (201 with
    `Location`, `ETag` and `Last-Modified`; 409 for a duplicate). Readings are append-only:
    PUT, PATCH and DELETE on a reading or on the readings collection return 405 with `Allow`.
    Registry: `POST /installations` (201 with `Location`, `ETag` and `Last-Modified`; 409 when
    the installation or meter id is taken), `PUT /installations/{installation-id}` (whole-document
    replacement, never a merge; PATCH returns 405) and `DELETE` (200, then 404; readings are kept),
    both honouring `If-Match` (a stale ETag → 412).
    `/installations` and the readings history are paginated and sortable; installations filter by
    jurisdiction and `status`, readings by time window and `min-power-kw`.
...

...

...

...

...
    and both summaries.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /installations:
    get:
      tags: [Installations]
      operationId: listInstallations
      summary: List installations in the caller's jurisdiction (paginated)
      description: |
        One page as `{count, next, previous, results}`, items in the plain shape (no last_reading).
        Filters combine (AND); every malformed filter or page value is reported in one 400, then the
        sort is checked. Links keep every filter and the sort. Ties are ordered by installation_id.
      parameters:
        - $ref: '#/components/parameters/ProvinceIdFilter'
        - $ref: '#/components/parameters/DistrictIdFilter'
        - $ref: '#/components/parameters/SubstationIdFilter'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/InstallationSort'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: One page of installations (plain shape, no last_reading)
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationPage' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400':
          description: INVALID_QUERY_PARAMETER (a malformed filter, status or page value; all reported at once) or INVALID_SORT_FIELD (reference lists installation_id, name, capacity_kw)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                invalidQueryParameter:
                  summary: Two malformed values reported together
                  value:
                    code: INVALID_QUERY_PARAMETER
                    message: One or more query parameters are invalid.
                    details:
                      - { field: district-id, location: query, issue: 'Must match DT-01, given once.', reference: null }
                      - { field: status, location: query, issue: 'Must be one of: active, inactive, given once.', reference: null }
                    more_info: /api-docs#error-codes
                invalidSortField:
                  summary: sort names a field installations cannot be sorted by
                  value:
                    code: INVALID_SORT_FIELD
                    message: The sort field is not supported.
                    details:
                      - { field: sort, location: query, issue: 'Use one allowed field, prefixed with - for descending, given once.', reference: 'installation_id, name, capacity_kw' }
                    more_info: /api-docs#error-codes
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [Installations]
      operationId: createInstallation
      summary: Register a new installation (registry admin)
      description: |
        The admin supplies the readable `installation_id`. `district_id` and `province_id` are copied
        from the substation by the server; sending them (or `last_reading`) is a 400.
        Registry-admin authentication is a placeholder for now (every request is accepted).
        Order of checks: 415 → 400 INVALID_JSON / 413 → 400 VALIDATION_FAILED → 400 UNKNOWN_SUBSTATION → 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InstallationCreate' }
      responses:
        '201':
          description: Created. The body is the new installation as the composite (last_reading is null).
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationComposite' }
        '400':
          description: VALIDATION_FAILED (one details entry per field), UNKNOWN_SUBSTATION (substation_id names no substation; 400 because the problem is a body field, not the URI), or INVALID_JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                validationFailed:
                  summary: A server-set field was sent and capacity is invalid
                  value:
                    code: VALIDATION_FAILED
                    message: The installation has 2 problems.
                    details:
                      - { field: district_id, location: body, issue: 'Not accepted: the server copies it from the substation.', reference: null }
                      - { field: capacity_kw, location: body, issue: Must be a number greater than 0., reference: null }
                    more_info: /api-docs#error-codes
                unknownSubstation:
                  summary: substation_id is well-formed but does not exist
                  value:
                    code: UNKNOWN_SUBSTATION
                    message: No grid substation with id SS-999.
                    details:
                      - { field: substation_id, location: body, issue: No grid substation with this id exists., reference: null }
                    more_info: /api-docs#error-codes
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '409':
          description: INSTALLATION_ID_TAKEN (reference is the existing installation) or METER_ID_TAKEN
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                installationIdTaken:
                  summary: installation_id is already registered
                  value:
                    code: INSTALLATION_ID_TAKEN
                    message: Installation INS-0001 is already registered.
                    details:
                      - { field: installation_id, location: body, issue: Already registered., reference: /installations/INS-0001 }
                    more_info: /api-docs#error-codes
                meterIdTaken:
                  summary: meter_id is fitted to another installation
                  value:
                    code: METER_ID_TAKEN
                    message: Meter MTR-000001 is already fitted to another installation.
                    details:
                      - { field: meter_id, location: body, issue: Already used by another installation., reference: null }
                    more_info: /api-docs#error-codes
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /installations/{installation-id}:
    parameters:
      - $ref: '#/components/parameters/InstallationId'
    get:
      tags: [Installations]
      operationId: getInstallation
      summary: Get an installation with its latest reading (composite)
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: The installation composite. last_reading is null if there are no readings yet.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationComposite' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/InstallationNotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    put:
      tags: [Installations]
      operationId: replaceInstallation
      summary: Replace a whole installation (registry admin)
      description: |
        Whole-document replacement. Every writable field must be sent; a missing one is a 400, never
        filled in from the stored version (no merge). PUT never creates: an unknown id is 404, decided
        before the body is read. `installation_id` may be left out; if sent it must equal the path
        (otherwise 400 ID_MISMATCH). `district_id` and `province_id` are recalculated from the
        substation. The meter's credential and `created_at` are kept; readings keep the jurisdiction
        they were recorded under. Registry-admin authentication is a placeholder for now.
        Optional `If-Match`: the ETag last received for this installation (from GET, POST or PUT). A stale one
        → 412 and nothing changes; checked after the 404 and before the body.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InstallationReplace' }
      responses:
        '200':
          description: Replaced. The body is the installation as the composite.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationComposite' }
        '400':
          description: VALIDATION_FAILED (including any missing field), ID_MISMATCH, UNKNOWN_SUBSTATION, or INVALID_JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                partialBody:
                  summary: A partial (merge-style) body is refused
                  value:
                    code: VALIDATION_FAILED
                    message: The installation has 8 problems.
                    details:
                      - { field: meter_id, location: body, issue: Required., reference: null }
                      - { field: substation_id, location: body, issue: Required., reference: null }
                    more_info: /api-docs#error-codes
                idMismatch:
                  summary: The body names a different installation than the URL
                  value:
                    code: ID_MISMATCH
                    message: The body's installation_id does not match the URL (INS-0001).
                    details:
                      - { field: installation_id, location: body, issue: 'Must equal INS-0001, or be left out.', reference: /installations/INS-0001 }
                    more_info: /api-docs#error-codes
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/InstallationNotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '409':
          description: METER_ID_TAKEN. The meter is fitted to another installation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: METER_ID_TAKEN, message: Meter MTR-000002 is already fitted to another installation., details: [ { field: meter_id, location: body, issue: Already used by another installation., reference: null } ], more_info: /api-docs#error-codes }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      tags: [Installations]
      operationId: patchInstallationNotAllowed
      summary: Not allowed. Use PUT with the whole installation (405)
      security: []
      responses:
        '405':
          description: METHOD_NOT_ALLOWED. Partial updates are not offered; PUT replaces the whole installation.
          headers:
            Allow:
              description: The methods this URI does support
              schema: { type: string, example: 'GET, HEAD, PUT, DELETE' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: METHOD_NOT_ALLOWED, message: 'Partial updates are not supported: use PUT with the whole installation. Allowed: GET, HEAD, PUT, DELETE.', details: [], more_info: /api-docs#error-codes }
    delete:
      tags: [Installations]
      operationId: deleteInstallation
      summary: Remove an installation (registry admin). Its readings are kept.
      description: |
        Returns 200 with the installation as it was just before removal; a second DELETE returns 404.
        There is no cascade: the installation's readings stay in the database as the record of what
        was generated, still counted under the district they were recorded in. Because their parent
        no longer exists, the scoped readings URIs then return 404.
        Registry-admin authentication is a placeholder for now.
        Optional `If-Match`: a stale ETag → 412 and nothing is deleted; checked after the 404.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Removed. The body is the installation as the composite, as it was just before removal. A second DELETE returns 404.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationComposite' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/InstallationNotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /installations/{installation-id}/last-known-reading:
    parameters:
      - $ref: '#/components/parameters/InstallationId'
    get:
      tags: [Readings]
      operationId: getLastKnownReading
      summary: The most recent reading of an installation (derived)
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: The latest reading
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reading' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: INSTALLATION_NOT_FOUND, or NO_READINGS_YET when the installation exists but has no readings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                noReadingsYet:
                  summary: The installation exists but has not reported yet
                  value: { code: NO_READINGS_YET, message: Installation INS-0030 has not reported any readings yet., details: [], more_info: /api-docs#error-codes }
                installationNotFound:
                  summary: No such installation
                  value: { code: INSTALLATION_NOT_FOUND, message: No installation with id INS-9999., details: [], more_info: /api-docs#error-codes }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /installations/{installation-id}/readings/{reading-id}:
    parameters:
      - $ref: '#/components/parameters/InstallationId'
      - $ref: '#/components/parameters/ReadingId'
    get:
      tags: [Readings]
      operationId: getReading
      summary: One reading (the target of a 201 Location)
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: The reading
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reading' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: INSTALLATION_NOT_FOUND when the parent is missing, or READING_NOT_FOUND when this installation has no such reading (including a real reading id that belongs to another installation)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                readingNotFound:
                  summary: No such reading under this installation
                  value: { code: READING_NOT_FOUND, message: Installation INS-0002 has no reading with id 66f2c1a9e4b0a1b2c3d4e5f6., details: [], more_info: /api-docs#error-codes }
                installationNotFound:
                  summary: The parent installation is missing
                  value: { code: INSTALLATION_NOT_FOUND, message: No installation with id INS-9999., details: [], more_info: /api-docs#error-codes }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    put:
      tags: [Readings]
      operationId: replaceReadingNotAllowed
      summary: Not allowed. A reading is never replaced (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedReading' }
    patch:
      tags: [Readings]
      operationId: patchReadingNotAllowed
      summary: Not allowed. A reading is never edited (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedReading' }
    delete:
      tags: [Readings]
      operationId: deleteReadingNotAllowed
      summary: Not allowed. A reading is never deleted (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedReading' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token. Scopes (installation-write, analyst-read-by-jurisdiction, asset-admin) are documented in Step 9.9.
  parameters:
...
    InstallationId:
      name: installation-id
      in: path
      required: true
      schema: { type: string, pattern: '^INS-[0-9]{4}$', example: INS-0001 }
    ReadingId:
      name: reading-id
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9a-f]{24}$', example: 66f2c1a9e4b0a1b2c3d4e5f6 }
    ProvinceIdFilter:
      name: province-id
      in: query
      description: Only members in this province. Combines with the other filters (all must match). Malformed or repeated → 400 INVALID_QUERY_PARAMETER. From the security phase it is also intersected with the caller's jurisdiction (outside it → 403).
      schema: { type: string, pattern: '^PV-[0-9]{2}$', example: PV-01 }
    DistrictIdFilter:
      name: district-id
      in: query
      description: Only members in this district. Combines with the other filters (all must match). Malformed or repeated → 400 INVALID_QUERY_PARAMETER. From the security phase it is also intersected with the caller's jurisdiction (outside it → 403).
      schema: { type: string, pattern: '^DT-[0-9]{2}$', example: DT-01 }
    SubstationIdFilter:
      name: substation-id
      in: query
      description: Only installations on this substation. Combines with the other filters (all must match). Malformed or repeated → 400 INVALID_QUERY_PARAMETER. From the security phase it is also intersected with the caller's jurisdiction (outside it → 403).
      schema: { type: string, pattern: '^SS-[0-9]{3}$', example: SS-001 }
    Status:
      name: status
      in: query
      description: Only installations with this status. Any other value, or the parameter given twice → 400.
      schema: { type: string, enum: [active, inactive] }
...
    InstallationSort:
      name: sort
      in: query
      description: One field, written as in the JSON; prefix with - for descending. Ties are ordered by installation_id. Anything else → 400 INVALID_SORT_FIELD.
      schema:
        type: string
        enum: [installation_id, -installation_id, name, -name, capacity_kw, -capacity_kw]
        default: installation_id
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page-size
      in: query
      description: Items per page. Above 100 is a 400, never silently reduced.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag(s) the client holds. A match returns 304.
      schema: { type: string, example: '"a3f9c2d1"' }
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      description: Used only when If-None-Match is absent. Not modified since this date returns 304.
      schema: { type: string, example: 'Tue, 06 Oct 2026 04:30:00 GMT' }
    IfMatch:
      name: If-Match
      in: header
      description: Optional. The ETag the client last received for this installation (from GET, POST or PUT), a comma-separated list of them, or * for any existing installation. Strong comparison, so W/ tags never match. A mismatch returns 412 and nothing changes.
      schema: { type: string, example: '"a3f9c2d1"' }
  headers:
...
    ETag:
      description: Strong validator, a quoted hash of the response body
      schema: { type: string, example: '"a3f9c2d1"' }
    LastModified:
      description: When the representation last changed. Sent on single resources only, never on a collection or page.
      schema: { type: string, example: 'Tue, 06 Oct 2026 04:30:04 GMT' }
    Location:
      description: URI of the created resource. It resolves with GET.
      schema: { type: string, example: /installations/INS-0001/readings/66f2c1a9e4b0a1b2c3d4e5f6 }
    Vary:
      schema: { type: string, example: 'Accept, Authorization' }
    CacheControl:
      schema: { type: string, example: 'private, no-cache' }
    WWWAuthenticate:
      schema: { type: string, example: 'Bearer realm="slsea-solar-api"' }
...
  responses:
    NotModified:
      description: The client's copy is current. No body. Last-Modified is present only for single resources.
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
        Last-Modified: { $ref: '#/components/headers/LastModified' }
        Vary: { $ref: '#/components/headers/Vary' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
...
    Unauthorized:
      description: Missing, invalid or expired token
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Valid token but not allowed (scope, installation or jurisdiction)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotAcceptable:
      description: The Accept header excludes application/json
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PayloadTooLarge:
      description: The request body is over 100 KB
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnsupportedMediaType:
      description: The request body is not application/json
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InstallationNotFound:
      description: No installation with this id; for a scoped resource, the parent is missing (INSTALLATION_NOT_FOUND)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: INSTALLATION_NOT_FOUND, message: No installation with id INS-9999., details: [], more_info: /api-docs#error-codes }
...
    ServiceUnavailable:
      description: The database cannot be reached (SERVICE_UNAVAILABLE)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: SERVICE_UNAVAILABLE, message: The database cannot be reached., details: [], more_info: /api-docs#error-codes }
  schemas:
...
    InstallationComposite:
      description: An installation plus exactly one nested latest reading (null if none). Never the history.
      allOf:
        - $ref: '#/components/schemas/Installation'
        - type: object
          required: [last_reading]
          properties:
            last_reading:
              # OpenAPI 3.0.3: nullable only takes effect when type is set in the same schema.
              description: The newest reading, or null when the installation has not reported yet.
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/Reading'
    InstallationReplace:
      description: Every writable field. Server-set fields (district_id, province_id, last_reading) are not accepted.
      type: object
      additionalProperties: false
      required: [name, meter_id, substation_id, capacity_kw, status, commissioned_at, address, latitude, longitude]
      properties:
        installation_id: { type: string, description: Optional; if present it must equal the path id, example: INS-0221 }
        name: { type: string, example: Nugegoda rooftop 4 }
        meter_id: { type: string, example: MTR-000221 }
        substation_id: { type: string, example: SS-003 }
        capacity_kw: { type: number, minimum: 0, exclusiveMinimum: true, example: 4.2 }
        status: { type: string, enum: [active, inactive], example: active }
        commissioned_at: { type: string, format: date-time, example: '2026-10-01T00:00:00Z' }
        address: { type: string, example: '45 High Level Road, Nugegoda' }
        latitude: { type: number, minimum: -90, maximum: 90, example: 6.8649 }
        longitude: { type: number, minimum: -180, maximum: 180, example: 79.8997 }
    InstallationCreate:
      description: The admin supplies the readable installation_id.
      allOf:
        - $ref: '#/components/schemas/InstallationReplace'
        - type: object
          required: [installation_id]
...
    InstallationPage:
      type: object
      required: [count, next, previous, results]
      properties:
        count: { type: integer, minimum: 0, example: 220 }
        next: { type: string, nullable: true, example: '/installations?page=2&page-size=50&sort=installation_id' }
        previous: { type: string, nullable: true, example: null }
        results:
          type: array
          items: { $ref: '#/components/schemas/Installation' }
...
    Error:
      type: object
      required: [code, message, details, more_info]
      properties:
        code: { type: string, description: Stable machine-readable code (see Error codes), example: VALIDATION_FAILED }
        message: { type: string, example: The reading has 1 invalid field. }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        more_info: { type: string, example: '/api-docs#error-codes' }